src/restserver/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.7
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.0
PyJWT==1.7.1
requests==2.20.0
six==1.12.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
